# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "JAC Rental Booking Service API",
        "description": "Standalone microservice handling all rental property booking operations for the JAC platform. Authentication uses Bearer tokens issued by the JAC monolith (Laravel Passport).",
        "contact": {
            "email": "dev@jac.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "L5_SWAGGER_CONST_HOST",
            "description": "Current Environment"
        }
    ],
    "paths": {
...
        "/api/v1/coupons/apply": {
            "post": {
                "tags": [
                    "Coupons"
                ],
                "summary": "Validate and apply a coupon code",
                "description": "Validates the coupon code against the booking amount and returns the discount details.",
                "operationId": "acd68b67b817c20655d076768252570b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code",
                                    "amount"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "example": "SAVE20"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 500
                                    },
                                    "org_id": {
                                        "type": "integer",
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Coupon applied successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "example": "SAVE20"
                                        },
                                        "discount_type": {
                                            "type": "string",
                                            "example": "percent"
                                        },
                                        "discount_value": {
                                            "type": "number",
                                            "example": 20
                                        },
                                        "discount_amount": {
                                            "type": "number",
                                            "example": 100
                                        },
                                        "original_amount": {
                                            "type": "number",
                                            "example": 500
                                        },
                                        "final_amount": {
                                            "type": "number",
                                            "example": 400
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid coupon"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/favourites": {
            "get": {
                "tags": [
                    "Favourites"
                ],
                "summary": "List user favourites",
                "description": "Returns all properties saved to the authenticated user favourites list.",
                "operationId": "4a819eb9aad19fd2eafd2d41652b5686",
                "responses": {
                    "200": {
                        "description": "List of favourited properties",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/FavouriteItem"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Favourites"
                ],
                "summary": "Save a property to favourites",
                "description": "Adds a property to the user favourites. Idempotent.",
                "operationId": "1e30f16b1e0afdf9f1d15c3889703e7b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "property_id"
                                ],
                                "properties": {
                                    "property_id": {
                                        "type": "integer",
                                        "example": 12
                                    },
                                    "unit_id": {
                                        "type": "integer",
                                        "example": 34
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Added to favourites"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/favourites/toggle": {
            "post": {
                "tags": [
                    "Favourites"
                ],
                "summary": "Toggle favourite status",
                "description": "Adds if not saved; removes if saved.",
                "operationId": "02677a2ad530a10ff26e094ad4d8fc94",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "property_id"
                                ],
                                "properties": {
                                    "property_id": {
                                        "type": "integer",
                                        "example": 12
                                    },
                                    "unit_id": {
                                        "type": "integer",
                                        "example": 34
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Toggle result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Added to favourites"
                                        },
                                        "is_favourite": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/favourites/{id}": {
            "delete": {
                "tags": [
                    "Favourites"
                ],
                "summary": "Remove a favourite by ID",
                "description": "Deletes a specific favourite record.",
                "operationId": "6e9a52ba6b0f69bc2498a8fff2a2b792",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Removed from favourites"
                    },
                    "404": {
                        "description": "Favourite not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "FavouriteItem": {
                "description": "A saved favourite entry",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 10
                    },
                    "property_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "unit_id": {
                        "type": "integer",
                        "example": 5,
                        "nullable": true
                    },
                    "property": {
                        "$ref": "#/components/schemas/PropertySummary"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "PropertySummary": {
                "description": "Compact property representation returned in list views",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Sea View Villa"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "city_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "type_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "avg_rating": {
                        "type": "number",
                        "format": "float",
                        "example": 4.5
                    },
                    "rating_count": {
                        "type": "integer",
                        "example": 12
                    },
                    "is_featured": {
                        "type": "boolean",
                        "example": true
                    },
                    "star_rating": {
                        "type": "integer",
                        "example": 4
                    },
                    "status": {
                        "type": "string",
                        "example": "active"
                    },
                    "logo": {
                        "type": "string",
                        "example": "properties/villa1.jpg",
                        "nullable": true
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Passport Bearer token issued by the JAC monolith. Obtain via the monolith /api/v1/auth/login endpoint.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}